Fix spacewalk-repo-sync resetting /var/satellite/rhn/modules permissions to 750
Permissions are set to 770 when creating the modules directory to
allow Tomcat to write to comps/modules files as needed by the Web UI.

A later code path was resetting the permissions to 750 for reposync on
existing channels. Update that code to also set 770 for the modules dir.

Adopt fileutils.setPermsPath for consistency with the rest of the code.

Orabug: 31973025

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>