LINUX-8154 - Update modules.yaml when adding modular packages to channel
When adding a package from one modular channel to another, update the
destination channel's modules.yaml file so that dnf is aware of the
new package.

The modulemd document will be added to the modules.yaml file if
required. Any packages listed in the modulesmd artifacts rpms list
that are not in the destination channel will be added to the channel.

A new entity domain to model the modulemd documents has been
created. I doesn't model the whole document, just the elements
required for this functionality. When copying modulemd documents
between files the raw stream is used to copy the whole document.

Taskomatic is used to asynchronously update the repomd. A new checksum
for modules.yaml is generated, along with new package and errata
lists. As such, the new repo state is only available to dnf once
Tasomatic has processed these changes.

This functionality is only available from the Web UI.

Ant build targets and RPMs spec file have been updated to use
SnakeYAML found in the Spacewalk repo to manipulate the modules.yaml
file.

Jira: LINUX-8154
Orabug: 31960970

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>

Fix spacewalk-repo-sync resetting /var/satellite/rhn/modules permissions to 750

Permissions are set to 770 when creating the modules directory to
allow Tomcat to write to comps/modules files as needed by the Web UI.

A later code path was resetting the permissions to 750 for reposync on
existing channels. Update that code to also set 770 for the modules dir.

Adopt fileutils.setPermsPath for consistency with the rest of the code.

Orabug: 31973025

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>