LINUX-9346 - Handle modules.yaml file in spacewalk-manage-channel-lifecycle
spacewalk-manage-channel-lifecycle does not copy the modules.yaml
files when manipulating lifecycle channels. This causes the created
channels to reference to original channel's modules.yaml file. If the
original modules.yaml file is updated then it will reference newer
modules, streams, and packages that are not in the lifecycle channels,
which will cause client dnf modularity operations to fail.

This patch updates spacewalk-manage-channel-lifecycle so that all
operations keep the lifecycle channel's modules.yaml file in sync with
the package state.

For archive, rollback and channel promotion operations with the clear
option set, the modules file is copied wholesale because it's faster
than computing the modules stanza delta. For channel promotions with a
package delta the module set is computed and missing stanzas are added
to the target channel's modules.yaml file.

ChannelSofwareHandler has had clear() and rollbackChannel() methods
added to the channel.software XMLRPC API Namespace to service this new
functionality. clear() has been added to ensure that removing errata
and packages, and clearing the modules.yaml are an atomic operation.
This also avoids exposing the dangerous modules file methods to the
public API. spacewalk-manage-channel-lifecycle uses the new clear()
method instead of individual removeErrata() and removePackages()
methods.

XMLRPC API channel.software.mergePackages() takes a new parameter to
determine if the modules.yaml file should be updated. The method was
overloaded to ensure existing customer scripts don't break.

Jira: LINUX-9346

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Alex Burmashev <alexander.burmashev@oracle.com>