Fix logic inversion introduced in e62caa33ec.
A logical not was removed when converting == to equals() calls.

This had the affect of causing unassociated packages to be added to a
channel when adding custom errata from another channel. The reported
symptom was that OL6 packages associated with an errata were added
when adding errata from an exclusively OL7 channel to another
exclusively OL7 channel.

OLM doesn't have the concept of a distro version specific channel, so
the distro version side affect is merely coincidence. It could just as
easily have been completely unrelated OL7 packages.

This also fixes another instance of logic inversion in SCAP rules
comparison. The symptom is unrelated to this bug, but the underlying
cause is the same and introduced in the same commit.

Orabug: 32613033

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>