Make spacewalk-debug copy symlink target instead of the symlink itself.
Commit 02ebe01be6e6768 altered spacewalk-debug's log copying cp
arguments to "work with symlink files". The end result of this was
that symlinks to the log files were copied rather than the files
themselves. This is useless to a support rep because the content is
missing.

The upstream commit message isn't clear enough to determine the
intention. At any rate Oracle wants the log files included, so this
patch implements that behaviour.

Orabug: 32514543

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>