Fix errata scheduled job link ID number formatting.
The "scheduled" link to a successfully created errata scheduled job on
RelevantErrata.do was broken because large action IDs were being
formatted with commas.

Spacewalk uses Struts ActionMessage combined with XLIFF to provide I8N
message templates. ActionMessage defaults to MessageFormat's best
guess approach when passed an Object. This causes what appeared to be
a Long to be printed with commas. When this happens to ID numbers the
link breaks.

Use Long's .toString() to override MessageFormat's formatting and
prevent commas being added.

Orabug: 32581542

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>