Don't throw exception syncing multi-arch repos.
When syncing a repo that contains multiple package architectures into
a channel, an exception was thrown when packages with architecture
that was incompatible with the channel's architecture were
encountered. This aborted the entire sync run.

The error encountered was:
ERROR: Package arch XXXX incompatible with channel YYYY

This change replaces the exception with a debug level message and
continues processing. The end result is that packages with compatible
architecture are added to the channel, and those with incompatible
architecture are ignored.

The Import Class hierarchy is polymorphic, and therefore the methods
that invoke the Import.run() still catch the exceptions to account for
other forms of Import Classes.

Orabug: 35280456

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>