Fix rhnreg_ks failure when libvirtd not running.
VM host detection in rhnreg_ks reports a warning if
python3-rhn-virtualization-host is installed, but libvirtd is not
running.

In Python 3 this produces a TypeError because sys.stdout.write() now
expects str, not bytes. Special utf8 handling was introduced for
RHEL5/6, so we now remove it and just hand the Python3 str directly.

Note that the warning is valid in the context of rhn-virtualization on
a system that is not a KVM host, and harmless in the context of
rhnreg_ks for which the bugs were logged. This fix is specifically to
resolve the TypeError.

Orabug: 35466991 (OLM rhn-client-tools)
Orabug: 30366521 (Distro rhn-client-tools)

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>