Improve slow "Systems Overview" tab Web UI performance.
The Systems Overview tab was slow due to a chain of views composed of
scalar correlated subqueries and analytical functions calculating
package and errata vulnerabilities on yum modules data.

Convert the chain of 3 nested views to materialized views, with
indexes on columns that are used as restriction clauses from Java.

rhnServerOverview couldn't be directly replaced with an MV because
it's a scalar correlated subquery, and MV's can't directly reference
those. Therefore, the original view is renamed
rhnServerOverView_Backing, and the new MV selects from that. This
allows the Java middleware to remain unchanged because the referenced
view name is the same.

The materialised view queries are too complicated to be eligible for
fast refresh or "real-time" functionality. Similarly, using on-commit
refresh produces significant DML overhead because the views reference
many of the frequently updated core tables.

Therefore, MV refresh runs from a trigger on the Taskomatic
rhnTaskoRun table so that refresh only happens when a Taskomatic
errata or YUM metadata job has completed. This happens as an
autonomous transaction so as to not delay Taskomatic.

MV functionality is proprietary to Oracle, and not supported in OL7's
version of PostgreSQL, and therefore PostgreSQL is unchanged. The DDL
scripts have been refactored to separate the proprietary DDL.

Orabug: 36719403

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Alex Burmashev <alexander.burmashev@oracle.com>